customer-service-api/pyproject.toml
[tool.poetry]
name = "thisevent-fiap-auth-service"
version = "0.1.0"
description = ""
authors = ["OscarSilvaOfficial <oscarkaka222@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-RESTful = "^0.3.9"
PyMySQL = "^1.0.2"
SQLAlchemy-serializer = "^1.4.1"
PyJWT = "^2.3.0"
SQLAlchemy-Utils = "^0.37.9"
requests = "^2.26.0"
psycopg2-binary = "^2.9.2"
pytest = "^6.2.5"
mysql-connector-python = "^8.0.27"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
